pyproject.toml
[project]
name = "devsnek"
version = "0.1.0"
description = "Python development web server with LetsEncrypt support and ASGI capabilities"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "certbot>=2.0.0",
    "acme>=2.0.0",
    "cryptography>=40.0.0",
    "josepy>=1.13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
flask = ["flask>=2.0.0"]
fastapi = ["fastapi>=0.100.0", "uvicorn>=0.22.0"]

[project.scripts]
devsnek = "devsnek.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["devsnek"]